provide a common implementation
